Simplify test fixture
The reason for the with statement was I wanted to make sure the file
was properly closed after use.

But in fact, we're just mocking the opening of the file here, and it
will be closed in the get_dossiers_legislatifs function, so we don't
need to close it a second time, and we can make it simpler.