Revert smooth scrolling as is not consistent
When switching from one response to the next one, the behavior of .scroll() is buggy because the content of the previous response is hidden and thus change the position of the header which serve as reference for smooth scrolling. If it is still within the window that’s OK but if not it leads to inconsistent positions with hidden pertinent parts so let’s leave it as is for now.