Fix updating of self-referential parent amendement relationship
In some cases, the relationship would not be persisted correctly.

When fetching amendements from Assemblée nationale, only the last
sous-amendement would have its parent set correctly.

This commits makes sure the relationship is configured correctly,
and that we issue a DB flush when setting the parent amendement.

I'm not sure I understand the exact mechanics of the problem, but
together these two changes fix the reported issue.