Force a page refresh in browser tests where we change the data
It looks like the browser uses a previously cached version of the page,
so forcing a refresh solves the issue for now.

This way we can still use a single browser instance for all tests.

Using a new browser instance for each test, with a function-scoped fixture,
would also solve the issue, but that makes the tests run 3-4x slower.

We should probably look into why the page is cached when it should not though... 🤔