Synchronuous articles fetching on lecture creation
Otherwise the async task is locked given that it shares the lock with async amendements fetching which leads to a 5 minutes delay between lecture creation and articles’ content retrieval.

The trade-off is to wait a bit on creation but articles retrieval is less time consuming (only one HTML page to scrape).