Refactor auth token expiration to make it easier to test
When the expiration is only managed by Redis, then we actually have to until the key
expires, making the test slow.

If we include the expiration timestamp in the data, then the app can double-check it
on retrieval, and we can simulate time moving forward in tests without actually waiting.